+++ sensors/urls.py
@@ -17,6 +17,9 @@ import sensors.views as views
 router = routers.SimpleRouter()
 router.register(r'sensors', views.SensorViewSet, basename='sensors')
 
+sensors_router = routers.NestedSimpleRouter(router, r'sensors', lookup='sensor')
+sensors_router.register(r'logs', views.LogViewSet, basename='logs')
+
 ###
 # URLs
 ###
@@ -26,5 +29,7 @@ urlpatterns = [
         views.webhook,
         name='webhook',
     ),
+    re_path(r'sensors/(?P<pk>\d+)/dashboard-data', views.DashboardDataAPIView.as_view(), name='dashboard-data'),
     re_path(r'^', include(router.urls)),
+    re_path(r'^', include(sensors_router.urls)),
 ]