+++ sensors/views.py
@@ -7,11 +7,13 @@ from django.utils import timezone
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_http_methods
 from rest_framework import status
+from rest_framework.generics import RetrieveAPIView
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
 import sensors.models as models
-from sensors.serializers import SensorSerializer
+from sensors.serializers import SensorSerializer, SensorLogSerializer, DashboardDataSerializer
 
 
 # Create your views here.
@@ -65,3 +67,34 @@ class SensorViewSet(ModelViewSet):
         last_90_days = timezone.now() - datetime.timedelta(days=90)
         logs = models.Log.objects.filter(timestamp__gte=last_90_days)
         return models.Sensor.objects.all().prefetch_related(Prefetch('logs', queryset=logs))
+
+    def retrieve(self, request, *args, **kwargs):
+        ret = super().retrieve(request, *args, **kwargs)
+        return ret
+
+
+class LogViewSet(ModelViewSet):
+    serializer_class = SensorLogSerializer
+    permission_classes = [IsAuthenticated, ]
+    pagination_class = None
+
+    def get_queryset(self):
+        last_90_days = timezone.now() - datetime.timedelta(days=90)
+        logs = models.Log.objects.filter(timestamp__gte=last_90_days)
+        return models.Sensor.objects.filter(
+            id=self.kwargs.get('sensor_pk')
+        ).prefetch_related(
+            Prefetch('logs', queryset=logs)
+        )
+
+
+class DashboardDataAPIView(RetrieveAPIView):
+    serializer_class = DashboardDataSerializer
+    permission_classes = [IsAuthenticated, ]
+    queryset = models.Sensor.objects.all()
+    pagination_class = None
+
+    def get(self, request, *args, **kwargs):
+        instance = self.get_object()
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)