+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [](https://github.com/arthurdedeus/lora-api/compare/v0.2.0...v) (2022-05-18)
+
+
+### Features
+
+* Create DashboardDataAPIView ([ef8370d](https://github.com/arthurdedeus/lora-api/commit/ef8370d4f096e3b346466408fbdb1995f9eda4ac))
+* Create LogViewSet ([62c6489](https://github.com/arthurdedeus/lora-api/commit/62c648902c388a608e0699e75b24c909fb516a3d))
+
+
+
 # [](https://github.com/arthurdedeus/lora-api/compare/v0.1.0...v) (2022-05-11)